На вторую проверку ТЗ№10

+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -7,9 +7,14 @@ import java.util.List;
 
 public interface FilmStorage {
     void create(Film film);
+
     void update(Film film);
+
     ArrayList<Film> getAll();
-    Film getById(int id);
-    void deleteById(int id);
+
+    Film getById(long id);
+
+    void deleteById(long id);
+
     List<Film> getBestFilms(int count);
 }
