+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -7,8 +7,12 @@ import java.util.ArrayList;
 
 public interface UserStorage {
     void create(User user);
+
     void update(User user);
+
     ArrayList<User> getAll();
-    User getById(int id);
-    void deleteById(int id);
+
+    User getById(long id);
+
+    void deleteById(long id);
 }
